完善上传设置，后台上传文件添加文件限制

+++ Atlass.Framework.Models/CmsSet/cms_upload_set.cs
@@ -71,6 +71,6 @@ namespace Atlass.Framework.Models
         /// <summary>
         /// 文件最大大小
         /// </summary>
-        public int max_file_size { get; set; } = 10;
+        public long max_file_size { get; set; } = 10;
     }
 }
